Apply an exception in the delete command

+++ HR.LeaveManagement.Application/Features/LeaveTypes/Handlers/Commands/DeleteLeaveTypeCommandHandler.cs
@@ -1,4 +1,5 @@
-﻿using HR.LeaveManagement.Application.Features.LeaveTypes.Requests.Commands;
+﻿using HR.LeaveManagement.Application.Exceptions;
+using HR.LeaveManagement.Application.Features.LeaveTypes.Requests.Commands;
 using HR.LeaveManagement.Application.Persistence.Contracts;
 using MediatR;
 
@@ -16,8 +17,11 @@ namespace HR.LeaveManagement.Application.Features.LeaveTypes.Handlers.Commands
         public async Task<Unit> Handle(DeleteLeaveTypeCommand request, CancellationToken cancellationToken)
         {
             var leaveType = await _leaveTypeRepository.GetAsync(request.Id);
-            await _leaveTypeRepository.DeleteAsync(leaveType);
 
+            if (leaveType is null)
+                throw new NotFoundException(nameof(leaveType), request.Id);
+
+            await _leaveTypeRepository.DeleteAsync(leaveType);
             return Unit.Value;
         }
     }
